Bug 28779: (QA follow-up) Even more specific message
For the detail page, the MARC view etc. we use the text:

The record you requested does not exist (488).

I think this is better than "not found", because that sounds
like it might be an issue with the search index, which we can
rule out in this case.

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>
(cherry picked from commit 970fcc3af48d8b613b45af5111fe4cab8b8a7f28)
Signed-off-by: Fridolin Somers <fridolin.somers@biblibre.com>